implemented stuff for reading and writing data

+++ src/common.hpp
@@ -5,4 +5,4 @@
 
 #define DEF_FUNCTION(NAME) def(xstr(NAME), NAME, NAME ## _doc)
 
-#define SCOPE_ATTR(NAME) scope().attr( #NAME ) = NAME
+#define SCOPE_ATTR(NAME) scope().attr( #NAME ) = (int)NAME
